fix(lms): fix borrow repository

+++ src/main/java/com/mooc/library_management/repository/BorrowRepository.java
@@ -3,4 +3,4 @@ package com.mooc.library_management.repository;
 import com.mooc.library_management.domain.Borrow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BorrowRepository extends JpaRepository<Borrow,Integer> { }
+public interface BorrowRepository extends JpaRepository<Borrow,Long> { }
